// src/app_modules/auth/Models/enums/token-type.enum.ts
export enum TokenType {

    ACCESS_TOKEN = "ACCESS_TOKEN",
    REFRESH_TOKEN = "REFRESH_TOKEN",
    WS_ACCESS_TOKEN = "WS_ACCESS_TOKEN",
    WS_REFRESH_TOKEN = "WS_REFRESH_TOKEN",
    PRE_AUTHORIZATION_TOKEN = "PRE_AUTHORIZATION_TOKEN",
    ACTIVATION_TOKEN = "ACTIVATION_TOKEN",
    PHONE_NUMBER_VERIFICATION_TOKEN = "PHONE_NUMBER_VERIFICATION_TOKEN",
    EMAIL_VERIFICATION_TOKEN = "EMAIL_VERIFICATION_TOKEN"


}